+++ backend/kirikiri/src/test/java/co/kirikiri/controller/helper/ControllerTestHelper.java
@@ -1,0 +1,28 @@
+package co.kirikiri.controller.helper;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
+
+import co.kirikiri.common.interceptor.AuthInterceptor;
+import co.kirikiri.common.resolver.MemberIdentifierArgumentResolver;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.test.mock.mockito.MockBean;
+
+public class ControllerTestHelper extends RestDocsHelper {
+
+    @MockBean
+    protected AuthInterceptor authInterceptor;
+
+    @MockBean
+    private MemberIdentifierArgumentResolver memberIdentifierArgumentResolver;
+
+    @BeforeEach
+    void setUp() {
+        when(authInterceptor.preHandle(any(), any(), any()))
+                .thenReturn(true);
+        when(memberIdentifierArgumentResolver.resolveArgument(any(), any(), any(), any()))
+                .thenReturn("Bearer Token");
+        when(memberIdentifierArgumentResolver.supportsParameter(any()))
+                .thenReturn(true);
+    }
+}
