+++ backend/kirikiri/src/test/java/co/kirikiri/integration/helper/IntegrationTest.java
@@ -1,20 +1,20 @@
 package co.kirikiri.integration.helper;
 
-import co.kirikiri.DatabaseCleaner;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.RestAssured;
+import java.io.UnsupportedEncodingException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Sql(value = "/delete.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
+//@Sql(value = "/delete.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
 public class IntegrationTest {
 
     protected static final String API_PREFIX = "/api";
@@ -36,5 +36,10 @@ public class IntegrationTest {
     @AfterEach
     void tearDown() {
         databaseCleaner.execute();
+    }
+
+    protected <T> T jsonToClass(final String responseBody, final TypeReference<T> typeReference)
+            throws JsonProcessingException, UnsupportedEncodingException {
+        return objectMapper.readValue(responseBody, typeReference);
     }
 }
