+++ backend/kirikiri/src/test/java/co/kirikiri/persistence/helper/RepositoryTest.java
@@ -1,6 +1,10 @@
 package co.kirikiri.persistence.helper;
 
 import co.kirikiri.common.config.JpaConfig;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
@@ -9,8 +13,10 @@ import org.springframework.test.context.TestConstructor.AutowireMode;
 
 @DataJpaTest
 @ActiveProfiles("test")
-@Import({TestEntityManagerConfig.class, JpaConfig.class})
+@Import({JpaConfig.class})
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
 @TestConstructor(autowireMode = AutowireMode.ALL)
-public class RepositoryTest {
+public @interface RepositoryTest {
 
 }
