+++ backend/kirikiri/src/test/resources/org/springframework/restdocs/templates/request-parameters.snippet
@@ -1,10 +1,10 @@
-    |===
-        |Parameter|Type|Description|Restrict|Required
-    {{#parameters}}
-        |{{#tableCellContent}}`+{{name}}+`{{/tableCellContent}}
-        |{{#tableCellContent}}`+{{type}}+`{{/tableCellContent}}
-        |{{#tableCellContent}}{{description}}{{/tableCellContent}}
-        |{{#tableCellContent}}{{^restrict}}제한 없음{{/restrict}}{{#restrict}}{{.}}{{/restrict}}{{/tableCellContent}}
-        |{{#tableCellContent}}{{^optional}}true{{/optional}}{{#optional}}false{{/optional}}{{/tableCellContent}}
-    {{/parameters}}
-        |===
+|===
+|Parameter|Type|Description|Restrict|Required
+{{#parameters}}
+|{{#tableCellContent}}`+{{name}}+`{{/tableCellContent}}
+|{{#tableCellContent}}`+{{type}}+`{{/tableCellContent}}
+|{{#tableCellContent}}{{description}}{{/tableCellContent}}
+|{{#tableCellContent}}{{^restrict}}제한 없음{{/restrict}}{{#restrict}}{{.}}{{/restrict}}{{/tableCellContent}}
+|{{#tableCellContent}}{{^optional}}true{{/optional}}{{#optional}}false{{/optional}}{{/tableCellContent}}
+{{/parameters}}
+|===
