+++ backend/kirikiri/src/test/resources/org/springframework/restdocs/templates/response-fields.snippet
@@ -1,8 +1,8 @@
-    |===
-        |Field|Type|Description
-    {{#fields}}
-        |{{#tableCellContent}}`+{{path}}+`{{/tableCellContent}}
-        |{{#tableCellContent}}`+{{type}}+`{{/tableCellContent}}
-        |{{#tableCellContent}}{{description}}{{/tableCellContent}}
-    {{/fields}}
-        |===
+|===
+|Field|Type|Description
+{{#fields}}
+|{{#tableCellContent}}`+{{path}}+`{{/tableCellContent}}
+|{{#tableCellContent}}`+{{type}}+`{{/tableCellContent}}
+|{{#tableCellContent}}{{description}}{{/tableCellContent}}
+{{/fields}}
+|===
